+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/handler/sink/DrainSink.java
@@ -1,10 +1,7 @@
 package com.dwarfeng.judge.impl.handler.sink;
 
-import com.dwarfeng.judge.impl.handler.Sink;
 import com.dwarfeng.judge.stack.bean.dto.SectionReport;
 import org.springframework.stereotype.Component;
-
-import java.util.Objects;
 
 /**
  * 简单的丢弃掉所有信息的水槽。
@@ -13,16 +10,22 @@ import java.util.Objects;
  * @since beta-1.0.0
  */
 @Component
-public class DrainSink implements Sink {
+public class DrainSink extends AbstractSink {
 
-    public static final String SUPPORT_TYPE = "drain";
+    public static final String SINK_TYPE = "drain";
 
-    @Override
-    public boolean supportType(String type) {
-        return Objects.equals(SUPPORT_TYPE, type);
+    public DrainSink() {
+        super(SINK_TYPE);
     }
 
     @Override
     public void sinkData(SectionReport sectionReport) {
+    }
+
+    @Override
+    public String toString() {
+        return "DrainSink{" +
+                "sinkType='" + sinkType + '\'' +
+                '}';
     }
 }